add greedy ai

+++ src/ts_interop.rs
@@ -6,7 +6,7 @@ use crate::piece::PieceName;
 
 #[derive(Serialize, Deserialize, TS)]
 #[ts(export, export_to = "pkg/types/V3.ts")]
-pub struct V3(#[ts(type = "[number, number,number]")] pub Vector3<i8>);
+pub struct V3(#[ts(type = "[number, number,number]")] pub Vector3<f32>);
 
 #[derive(Serialize, Deserialize, TS)]
 #[serde(tag = "type", content = "data")]
@@ -29,9 +29,9 @@ pub enum RotationAxis {
     Y,
 }
 
-#[derive(Serialize, Deserialize, TS, Default)]
+#[derive(Serialize, Deserialize, TS, Default, Clone, Debug)]
 #[ts(export, export_to = "pkg/types/Score.ts")]
 pub struct Score {
-    pub p1: usize,
-    pub p2: usize,
+    pub p1: i8,
+    pub p2: i8,
 }
